Create Roads and Libraries HackerRank.cpp
The Ruler of HackerLand believes that every citizen of the country should have access to a library. Unfortunately, HackerLand was hit by a tornado that destroyed all of its libraries and obstructed its roads! As you are the greatest programmer of HackerLand, the ruler wants your help to repair the roads and build some new libraries efficiently.

HackerLand has 𝖓 cities numbered from 𝟏 to 𝖓. The cities are connected by 𝖒 bidirectional roads. A citizen has access to a library if:

⇒ Their city contains a library.
⇒ They can travel by road from their city to a city containing a library.
The following figure is a sample map of HackerLand where the dotted lines denote obstructed roads:


The cost of repairing any road is 𝙘_𝙧𝙤𝙖𝙙𝙨 dollars, and the cost to build a library in any city is  dollars. If in the above example 𝙘_𝙧𝙤𝙖𝙙𝙨=2 and 𝙘_𝙡𝙞𝙗=3 , we would build 𝟓 roads at a cost of 𝟓 𝐱 𝟐 = 𝟏𝟎 and 𝟐 libraries for a cost of 𝟔. We don't need to rebuild one of the roads in the cycle 𝟭=>𝟮=>𝟯=>𝟭 .

You are given 𝙦 queries, where each query consists of a map of HackerLand and value of 𝙘_𝙧𝙤𝙖𝙙𝙨  and 𝙘_𝙡𝙞𝙗. For each query, find the minimum cost of making libraries accessible to all the citizens and print it on a new line.

𝐅𝐮𝐧𝐜𝐭𝐢𝐨𝐧 𝐃𝐞𝐬𝐜𝐫𝐢𝐩𝐭𝐢𝐨𝐧

Complete the function roadsAndLibraries in the editor below. It must return the minimal cost of providing libraries to all, as an integer.

roadsAndLibraries has the following parameters:

⇒ n: integer, the number of cities
⇒ c_lib: integer, the cost to build a library
⇒ c_road: integer, the cost to repair a road
⇒ cities: 2D array of integers where each  contains two integers that represent cities 
   connected by an obstructed road .

𝐈𝐧𝐩𝐮𝐭 𝐅𝐨𝐫𝐦𝐚𝐭

The first line contains a single integer 𝙦, that denotes the number of queries.

The subsequent lines describe each query in the following format:
⇒ The first line contains four space-separated integers that describe the respective values of 𝙣, 𝙢, 𝙘_𝙡𝙞𝙗 and 𝙘_𝙧𝙤𝙖𝙙 , the number of cities, number of roads, cost of a library and cost of a road.
⇒ Each of the next 𝙢 lines contains two space-separated integers, 𝒖[𝒊] and 𝒗[𝒊], that describe a bidirectional road that connects cities 𝒖[𝒊] and 𝒗[𝒊].

𝗢𝘂𝘁𝗽𝘂𝘁 𝗙𝗼𝗿𝗺𝗮𝘁

For each query, print an integer that denotes the minimum cost to make libraries accessible to all the citizens on a new line.

𝐒𝐚𝐦𝐩𝐥𝐞 𝐈𝐧𝐩𝐮𝐭

2
3 3 2 1
1 2
3 1
2 3
6 6 2 5
1 3
3 4
2 4
1 2
2 3
5 6

𝐒𝐚𝐦𝐩𝐥𝐞 𝐎𝐮𝐭𝐩𝐮𝐭

4
12