Find the nearest clone Hackerrank
In this challenge, there is a connected undirected graph where each of the nodes is a color. Given a color, find the shortest path connecting any two nodes of that color. Each edge has a weight of 𝟏. If there is not a pair or if the color is not found, print -𝟏.

For example, given 𝐠𝐫𝐚𝐩𝐡_𝐧𝐨𝐝𝐞𝐬 = 𝟓, and 4 edges 𝐠_𝐟𝐫𝐨𝐦 = [𝟏,𝟐,𝟐,𝟑] and 𝐠_𝐭𝐨 = [𝟐,𝟑,𝟒,𝟓] and colors for each node are 𝐚𝐫𝐫 = [𝟏,𝟐,𝟑,𝟏,𝟑] we can draw the following graph:

(𝟭/𝟭)-----(𝟮/𝟮)----(𝟯/𝟯)-----(𝟱/𝟯)
                 |
                 |
              (𝟰/𝟭)

Each of the nodes is labeled [node]/[color] and is colored appropriately. If we want the shortest path between color 𝟑, blue, we see there is a direct path between nodes  and . For green, color 𝟏, we see the path length 𝟐 from 𝟏->𝟐->𝟒. There is no pair for node 𝟒 having color 𝟐, red.

𝑭𝒖𝒏𝒄𝒕𝒊𝒐𝒏 𝑫𝒆𝒔𝒄𝒓𝒊𝒑𝒕𝒊𝒐𝒏
Complete the findShortest function in the editor below. It should return an integer representing the length of the shortest path between two nodes of the same color, or  if it is not possible.

findShortest has the following parameter(s):

=> g_nodes: an integer, the number of nodes
=> g_from: an array of integers, the start nodes for each edge
=> g_to: an array of integers, the end nodes for each edge
=> ids: an array of integers, the color id per node
=> val: an integer, the id of the color to match

𝑰𝒏𝒑𝒖𝒕 𝑭𝒐𝒓𝒎𝒂𝒕
The first line contains two space-separated integers 𝐧 and 𝐦, the number of nodes and edges in the graph.
Each of the next 𝐦 lines contains two space-separated integers 𝐠_𝐟𝐫𝐨𝐦[𝐢] and 𝐠_𝐭𝐨[𝐢], the nodes connected by an edge.
The next line contains 𝐧 space-seperated integers, 𝐢𝐝𝐬[𝐢], representing the color id of each node from 𝟏 to 𝐧.
The last line contains the id of the color to analyze.
𝐍𝐨𝐭𝐞: The nodes are indexed from 𝟏 to 𝐧.


𝑪𝒐𝒏𝒔𝒕𝒓𝒂𝒊𝒏𝒕𝒔:
𝟏<=𝐧<=𝟏𝟎^𝟔
𝟏<=𝐦<=𝟏𝟎^𝟔
𝟏<=𝐢𝐝𝐬[𝐢]<=𝟏𝟎^𝟖


𝑶𝒖𝒕𝒑𝒖𝒕 𝑭𝒐𝒓𝒎𝒂𝒕
Print the single integer representing the smallest path length or -𝟭.


𝑺𝒂𝒎𝒑𝒍𝒆 𝑰𝒏𝒑𝒖𝒕 
4 3
1 2
1 3
4 2
1 2 1 1 
1
𝑺𝒂𝒎𝒑𝒍𝒆 𝑶𝒖𝒕𝒑𝒖𝒕 
1 

𝑺𝒂𝒎𝒑𝒍𝒆 𝑰𝒏𝒑𝒖𝒕 
4 3
1 2
1 3
4 2
1 2 3 4
2
𝑺𝒂𝒎𝒑𝒍𝒆 𝑶𝒖𝒕𝒑𝒖𝒕 
-1


𝑺𝒂𝒎𝒑𝒍𝒆 𝑰𝒏𝒑𝒖𝒕 
5 4
1 2
1 3
2 4
3 5
1 2 3 3 2
2
𝑺𝒂𝒎𝒑𝒍𝒆 𝑶𝒖𝒕𝒑𝒖𝒕 
3